Type the finite Number constants as float64 so MAX_SAFE_INTEGER lowers to a float64 (#719)
Number.MAX_SAFE_INTEGER and Number.MIN_SAFE_INTEGER were written as bare integer
literals, so an untyped integer constant defaulted to int in a := binding. A
const bound to Number.MAX_SAFE_INTEGER then refused to flow into a float64
parameter (value.Some[float64], the Duration field builders), failing the Go
build on a cluster of Temporal cases. Every JS Number is a double, so type all
five finite Number constants float64. The two internal Go range checks in the
goimport bridge convert their int64/uint64 operand to compare against the
now-float64 bound; the check is unchanged in meaning.

+++ pkg/goimport/bridge/bridge_test.go
@@ -29,7 +29,7 @@ func TestStringFromGoIsBentoString(t *testing.T) {
 }
 
 func TestInt64ToNumberInRange(t *testing.T) {
-	for _, n := range []int64{0, 1, -1, value.NumberMaxSafeInteger, value.NumberMinSafeInteger} {
+	for _, n := range []int64{0, 1, -1, int64(value.NumberMaxSafeInteger), int64(value.NumberMinSafeInteger)} {
 		if got := Int64ToNumber(n); got != float64(n) {
 			t.Errorf("Int64ToNumber(%d) = %v, want %v", n, got, float64(n))
 		}
@@ -37,7 +37,7 @@ func TestInt64ToNumberInRange(t *testing.T) {
 }
 
 func TestInt64ToNumberOutOfRangeRaises(t *testing.T) {
-	for _, n := range []int64{value.NumberMaxSafeInteger + 1, value.NumberMinSafeInteger - 1} {
+	for _, n := range []int64{int64(value.NumberMaxSafeInteger) + 1, int64(value.NumberMinSafeInteger) - 1} {
 		func() {
 			defer func() {
 				r := recover()
@@ -62,7 +62,7 @@ func TestUint64ToNumberOutOfRangeRaises(t *testing.T) {
 			t.Errorf("raised %T, want RangeError", r)
 		}
 	}()
-	Uint64ToNumber(value.NumberMaxSafeInteger + 1)
+	Uint64ToNumber(uint64(value.NumberMaxSafeInteger) + 1)
 }
 
 func TestMustReturnsValueWhenNoError(t *testing.T) {
